requirements.txt
tornado==6.2
celery==5.2.7
redis==4.3.4
python-dotenv